backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
sentence-transformers
faiss-cpu
langchain-community
pdfplumber
